refactor: extract workspace artifact cleanup
